Update to fileserver format
