--- Portfolio NodeJS/education.html
@@ -1,86 +1,0 @@
-<!DOCTYPE HTML>
-
-<html>
-<head>
-    <link rel="stylesheet" type="text/css" href="/css/0generic.css">
-    <link rel="stylesheet" type="text/css" href="/css/1titlebar.css">
-    <link rel="stylesheet" type="text/css" href="/css/2cards.css">
-    <meta name="viewport" content="width=device-width">
-    <title>
-        Education
-    </title>
-</head>
-
-<body class="EntirePage">
-    <div class="TitleBar">
-        <b class="Title">Benz Chong's Portfolio</b>
-        <p></p>
-        <button class="MenuButton" onclick="location.href='/'" type="button"><b>▢ Homepage</b></button>
-        <button class="MenuButtonSelected" onclick="location.href='education'" type="button"><b>🖉 Education</b></button>
-        <button class="MenuButton" onclick="location.href='achievements'" type="button"><b>🎖 Achievements</b></button>
-        <button class="MenuButton" onclick="location.href='experience'" type="button"><b>⚙ Experience</b></button>
-    </div>
-    <p class="Separator"></p>
-
-    <div class="Card">
-        <text class="CardTitle">Highest Qualification: GCE 'O' Levels Equivalent</text>
-        <p class="CardContent">Mean Subject Grade (MSG): 1.00</p>
-        <img src="https://drive.google.com/uc?export=download&id=1D1Sv48F2VyVM6C5Fla2aLocIt5Rv6SLr" class="TinyIcon">
-        <text class="CardSubheader">Secondary 4 Integrated Programme Subjects</text>
-        <p class="CardContent">A1 | English Language</p>
-        <p class="CardContent">A1 | Integrated Mathematics</p>
-        <p class="CardContent">A1 | Additional Mathematics</p>
-        <p class="CardContent">A1 | Physics</p>
-        <p class="CardContent">A1 | Chemistry</p>
-        <p class="CardContent">A1 | Biology</p>
-        <p class="CardContent">A1 | Social Studies</p>
-        <p class="CardContent">A1 | Socio-Economic & China Studies</p>
-        <p class="CardContent">A1 | Project Work</p>
-        <img src="https://drive.google.com/uc?export=download&id=1-Ca_trY0oqS5gzcWsKY8lNwR2ifKo2hB" class="TinyIcon">
-        <text class="CardSubheader">Singapore-Cambridge GCE 'O' Level Subjects</text>
-        <p class="CardContent">A1 | Higher Chinese</p>
-    </div>
-
-    <div class="Card">
-        <img src="https://drive.google.com/uc?export=download&id=1qQf8tW_zBp_RvqW2F73By33opdOtc54X" class="CardIcon">
-        <br>
-        <text class="CardTitle">Hwa Chong Institution (College)&nbsp |&nbsp 2023-2024</text>
-        <p class="CardContent">Enrolled in the Integrated Programme in Hwa Chong Institution from 2019 to present.</p>
-    </div>
-
-    <div class="Card">
-        <img src="https://drive.google.com/uc?export=download&id=1RnJurCDraEtWH4TReyHe-SF4H2eWwjor" class="CardIcon">
-        <br>
-        <text class="CardTitle">Hwa Chong Institution (High School)&nbsp |&nbsp 2019-2022</text>
-        <p class="CardContent">Enrolled in the Integrated Programme in Hwa Chong Institution from 2019 to present.</p>
-        <p class="CardContent">Selected for the Science & Mathematics Talent Programme from 2021 to 2022.</p>
-        <p class="CardContent">Achieved an MSG of 1.10 in 2022.</p>
-    </div>
-
-    <div class="Card">
-        <img src="https://drive.google.com/uc?export=download&id=1VaFiFtQSXgIpX0Ug4CVK_4HkiH0A2rRa" class="CardIcon">
-        <br>
-        <text class="CardTitle">Nanyang Primary School&nbsp |&nbsp 2016-2018</text>
-        <p class="CardContent">Enrolled in the Gifted Education Programme in Nanyang Primary School from 2016 to 2018.</p>
-        <p class="CardContent">Achieved a PSLE T-score of 275 (i.e. AL4) in 2018.</p>
-    </div>
-
-    <br>
-    <div class="CardMega">Contact Information</div>
-    <div class="Card">
-        <text class="CardTitle">Get in touch</text>
-        <p class="CardContent">Please feel free to contact me via any of the following avenues:</p>
-        <p class="CardContent">(Preferred) Email Address: 191214n@student.hci.edu.sg</p>
-        <p class="CardContent">(Alternative 1) Mobile Number: +65 8692 8108 (Republic of Singapore)</p>
-        <p class="CardContent">(Alternative 2) WhatsApp: +65 8692 8108</p>
-    </div>
-
-    <div class="Card">
-        <p class="CardContent">© 2023 Benz Chong Tze Jun. All rights reserved.</p>
-        <p class="CardContent">All elements of this website are self-created.
-        No copyright claims have been infringed on.
-        Background image is royalty free, courtesy of Vecteezy.</p>
-    </div>
-
-    <button class="SnapToTop" onclick="location.href='education'" type="button">Ʌ</button>
-</body>
