--- Portfolio NodeJS/public/css/0generic.css
@@ -1,34 +1,0 @@
-.EntirePage {
-    background-image: url('https://drive.google.com/uc?export=download&id=1xsgA4z4Qfg9vHBeOzdv_eKn7H7cCSNHs');
-    background-repeat: no-repeat;
-    background-attachment: fixed;
-    margin: 0;
-}
-.Separator {
-    font-size: 30px;
-}
-
-
-
-.SnapToTop {
-    background-color: white;
-    font: bold 20px Sans-Serif;
-    color: #3F3F3F;
-    position: fixed;
-    bottom: 0;
-    right: 0;
-    border-radius: 50px;
-
-    border: none;
-    box-shadow: 0px 0px 10px 10px rgba(200,200,200,0.3);
-    margin: 30px;
-    padding: 10px;
-
-    transition-duration: 0.4s;
-}
-.SnapToTop:hover{
-    background-color: #3F3F3F;
-    color: white;
-    cursor: pointer;
-    scale: 1.35;
-}
