--- Portfolio NodeJS/public/css/1titlebar.css
@@ -1,98 +1,0 @@
-.TitleBar {
-    background: rgba(220, 220, 225,0.6);
-    backdrop-filter: blur(3px);
-    text-align: center;
-    padding-top: 15px;
-    padding-bottom: 15px;
-    box-shadow: 0px 0px 10px 10px rgba(200,200,200,0.8);
-
-    position: -webkit-sticky;
-    position: sticky;
-    top: 0;
-
-    z-index: 1;
-}
-.Title {
-    padding-top: 12px;
-    padding-bottom: 12px;
-    padding-right: 40px;
-    padding-left: 40px;
-    display: inline-block;
-    border-radius: 50px;
-
-    background-color: white;
-    box-shadow: 0px 0px 10px 10px rgba(200,200,200,0.6);
-
-    font-family: Sans-Serif;
-    font-size: 24px;
-    color: #3F3F3F;
-    text-align: center;
-
-    justify-content: center;
-
-    transition-duration: 0.4s;
-}
-.Title:hover{
-    scale: 1.05
-}
-.MenuButton {
-    padding-top: 10px;
-    padding-bottom: 10px;
-    padding-right: 20px;
-    padding-left: 20px;
-    display: inline-block;
-    border-radius: 50px;
-
-    background-color: white;
-    border: none;
-    box-shadow: 0px 0px 5px 5px rgba(200,200,200,0.6);
-    margin-left: 5px;
-    margin-right: 5px;
-
-    font-family: Sans-Serif;
-    font-size: 16px;
-    color: #3F3F3F;
-    text-align: center;
-
-    justify-content: center;
-
-    transition-duration: 0.4s;
-}
-.MenuButtonSelected{
-    padding-top: 10px;
-    padding-bottom: 10px;
-    padding-right: 20px;
-    padding-left: 20px;
-    display: inline-block;
-    border-radius: 50px;
-
-    background-color: white;
-    border: none;
-    box-shadow: 0px 0px 5px 5px rgba(200,200,200,0.6);
-    margin-left: 5px;
-    margin-right: 5px;
-
-    font-family: Sans-Serif;
-    font-size: 16px;
-    color: #3F3F3F;
-    text-align: center;
-
-    justify-content: center;
-
-    transition-duration: 0.4s;
-
-    background-color: #3F3F3F;
-    color: white;
-}
-.MenuButton:hover{
-    background-color: #3F3F3F;
-    color: white;
-    cursor: pointer;
-    scale: 1.15;
-}
-.MenuButtonSelected:hover{
-    background-color: #3F3F3F;
-    color: white;
-    cursor: pointer;
-    scale: 1.15;
-}
