--- Portfolio NodeJS/public/css/2cards.css
@@ -1,73 +1,0 @@
-.Card {
-    background-color: rgba(245, 245, 245, 0.8);
-    backdrop-filter: blur(3px);
-    padding: 20px;
-    box-shadow: 0px 0px 10px 10px rgba(200,200,200,0.3);
-    margin-left: auto;
-    margin-right: auto;
-    margin-top: 30px;
-    margin-bottom: 30px;
-    width: 80%;
-    border-radius: 20px;
-}
-.CardMega{
-    background-color: rgba(245, 245, 245, 0.8);
-    backdrop-filter: blur(5px);
-    padding: 20px;
-    box-shadow: 0px 0px 10px 10px rgba(200,200,200,0.3);
-    width: 800px;
-    margin-left: auto;
-    margin-right: auto;
-    margin-top: 30px;
-    margin-bottom: 30px;
-    width: 90%;
-    border-radius: 20px;
-
-    font: bold 30px Sans-Serif;
-    color: #3F3F3F;
-    text-align: center;
-
-    z-index: 1
-}
-.CardIcon {
-    height: 100px;
-    max-width: 100%;
-}
-.TinyIcon {
-    height: 15px;
-}
-.CardTitle {
-    font: bold 22px Sans-Serif;
-    color: #3F3F3F
-}
-.CardSubheader{
-    font: bold 16px Sans-Serif;
-    text-decoration: underline;
-    color: #3F3F3F
-}
-.CardContent {
-    font: bold 14px Sans-Serif;
-    color: #3F3F3F
-}
-.CardImage {
-    display: block;
-    margin-left: auto;
-    margin-right: auto;
-    margin-top: 20px;
-    margin-bottom: 20px;
-    width: 60%;
-    max-width: 250px;
-    max-height: 250px;
-    box-shadow: 0px 0px 10px 10px rgba(200,200,200,0.3);
-    border-radius: 20px;
-}
-.CardImageCaption {
-    text-align: center;
-    font: bold 20px Sans-Serif;
-    color: #3F3F3F;
-}
-.CardImageCaption2 {
-    text-align: center;
-    font: bold 16px Sans-Serif;
-    color: #3F3F3F;
-}
