+++ achievements.html
@@ -1,0 +1,99 @@
+<!DOCTYPE HTML>
+
+<html>
+<head>
+    <link rel="stylesheet" type="text/css" href="/css/0generic.css">
+    <link rel="stylesheet" type="text/css" href="/css/1titlebar.css">
+    <link rel="stylesheet" type="text/css" href="/css/2cards.css">
+    <meta name="viewport" content="width=device-width">
+    <title>
+        Achievements
+    </title>
+</head>
+
+<body class="EntirePage">
+    <div class="TitleBar">
+        <b class="Title">Benz Chong's Portfolio</b>
+        <p></p>
+        <button class="MenuButton" onclick="location.href='/'" type="button"><b>▢ Homepage</b></button>
+        <button class="MenuButton" onclick="location.href='education'" type="button"><b>🖉 Education</b></button>
+        <button class="MenuButtonSelected" onclick="location.href='achievements'" type="button"><b>🎖 Achievements</b></button>
+        <button class="MenuButton" onclick="location.href='experience'" type="button"><b>⚙ Experience</b></button>
+    </div>
+    <p class="Separator"></p>
+
+    <div class="CardMega">Academics</div>
+
+    <div class="Card">
+        <text class="CardTitle">Distinction in International Chemistry Quiz 2022 (Year 10)</text>
+        <p class="CardContent">Awarded Distinction in the International Chemistry Quiz 2022 that tests candidates on advanced Chemistry knowledge and higher order thinking</p>
+    </div>
+
+    <div class="Card">
+        <text class="CardTitle">Lee Hsien Loong All-Round Excellence Award</text>
+        <p class="CardContent">Awarded the Lee Hsien Loong All-Round Excellence Award that recognises students for commitment to academic excellence</p>
+    </div>
+
+    <div class="CardMega">Research & Innovation</div>
+
+    <div class="Card" onclick="location.href='experience'" type="button">
+        <text class="CardTitle">High Distinction in Project Work (Secondary 4)</text>
+        <p class="CardContent">Awarded High Distinction in Project Work for leading a team in research into the effect of herbs on glucose-treated Caenorhabditis elegans</p>
+        <p class="CardContent">Invited to present at the international Science & Technology Symposium for Youth</p>
+        <p class="CardContent">For more information, please click.</p>
+    </div>
+
+    <div class="Card">
+        <text class="CardTitle">High Distinction in Project Work (Secondary 3)</text>
+        <p class="CardContent">Awarded High Distinction in Project Work for research into the effect of herbs on glucose toxicity in Caenorhabditis elegans (proof of concept for above Secondary 4 project)</p>
+    </div>
+
+    <div class="Card">
+        <text class="CardTitle">High Distinction in Project Work (Secondary 2)</text>
+        <p class="CardContent">Awarded High Distinction in Project Work for leading a team in innovation and invention of an amplifier for front-facing speakers of smartphones</p>
+    </div>
+
+    <div class="Card">
+        <text class="CardTitle">High Distinction in Project Work (Secondary 1)</text>
+        <p class="CardContent">Awarded High Distinction in Project Work for leading a team in innovation and invention of a universal and ergonomic pen light for writing in the dark</p>
+    </div>
+
+    <div class="Card" onclick="location.href='experience#monkey'" type="button">
+        <text class="CardTitle">Gold Award in Individualised Research Study</text>
+        <p class="CardContent">Awarded the Gold Award in Individualised Research Study for research into the behaviour of long-tailed macaques in Singapore</p>
+        <p class="CardContent">Invited to present at the Roots & Shoots Conference organized by the Jane Goodall Institute (Singapore)</p>
+        <p class="CardContent">For more information, please click.</p>
+    </div>
+
+    <div class="CardMega">The Arts</div>
+
+    <div class="Card">
+        <text class="CardTitle">Distinction in Grade 8 Piano</text>
+        <p class="CardContent">Achieved Distinction in Grade 8 Piano, graded by the ABRSM</p>
+    </div>
+
+    <div class="Card">
+        <text class="CardTitle">Platinum Award in the 5th Singapore Performers’ Festival</text>
+        <p class="CardContent">Achieved the highest award, Platinum, at the 5th Singapore Performers’ Festival competition</p>
+        <p class="CardContent">Received the Pianomania Scholarship</p>
+    </div>
+
+    <br>
+    <div class="CardMega">Contact Information</div>
+    <div class="Card">
+        <text class="CardTitle">Get in touch</text>
+        <p class="CardContent">Please feel free to contact me via any of the following avenues:</p>
+        <p class="CardContent">(Preferred) Email Address: 191214n@student.hci.edu.sg</p>
+        <p class="CardContent">(Alternative 1) Mobile Number: +65 8692 8108 (Republic of Singapore)</p>
+        <p class="CardContent">(Alternative 2) WhatsApp: +65 8692 8108</p>
+    </div>
+
+    <div class="Card">
+        <p class="CardContent">© 2023 Benz Chong Tze Jun. All rights reserved.</p>
+        <p class="CardContent">All elements of this website are self-created.
+        No copyright claims have been infringed on.
+        Background image is royalty free, courtesy of Vecteezy.</p>
+    </div>
+
+    <button class="SnapToTop" onclick="location.href='achievements'" type="button">Ʌ</button>
+</body>
